requirements.txt
fastapi
uvicorn
python-dotenv
twilio
psycopg2-binary

